Add commented out version using updated cgets in cc65
This provides less code, but the upstream cc65 change
was only introduced 2/11/2025, so users will have to
compile using HEAD cc65 after that date.
Also it adds 120 bytes to the overall size compared
to the handcrafted version, and finally, it does not
support continuous editing until CR is hit, instead
as per the cgets standard, it exits when user enters
enough characters to fill buffer, which unfortunately
means no chance to edit the string if you are on the
last character.